Update page vote model.
