+++ deepwell/src/models/mod.rs
@@ -13,8 +13,8 @@ pub mod page_connection;
 pub mod page_connection_missing;
 pub mod page_link;
 pub mod page_parent;
-pub mod page_rate_vote;
 pub mod page_revision;
+pub mod page_vote;
 pub mod sea_orm_active_enums;
 pub mod settings;
 pub mod site;
