+++ deepwell/src/models/page.rs
@@ -1,4 +1,4 @@
-//! SeaORM Entity. Generated by sea-orm-codegen 0.5.0
+//! SeaORM Entity. Generated by sea-orm-codegen 0.6.0
 
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
@@ -46,14 +46,14 @@ pub enum Relation {
     Site,
     #[sea_orm(has_many = "super::file::Entity")]
     File,
-    #[sea_orm(has_many = "super::page_rate_vote::Entity")]
-    PageRateVote,
+    #[sea_orm(has_many = "super::page_revision::Entity")]
+    PageRevision,
+    #[sea_orm(has_many = "super::page_vote::Entity")]
+    PageVote,
     #[sea_orm(has_many = "super::page_link::Entity")]
     PageLink,
     #[sea_orm(has_many = "super::page_connection_missing::Entity")]
     PageConnectionMissing,
-    #[sea_orm(has_many = "super::page_revision::Entity")]
-    PageRevision,
 }
 
 impl Related<super::forum_thread::Entity> for Entity {
@@ -80,9 +80,15 @@ impl Related<super::file::Entity> for Entity {
     }
 }
 
-impl Related<super::page_rate_vote::Entity> for Entity {
+impl Related<super::page_revision::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::PageRateVote.def()
+        Relation::PageRevision.def()
+    }
+}
+
+impl Related<super::page_vote::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::PageVote.def()
     }
 }
 
@@ -95,12 +101,6 @@ impl Related<super::page_link::Entity> for Entity {
 impl Related<super::page_connection_missing::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::PageConnectionMissing.def()
-    }
-}
-
-impl Related<super::page_revision::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::PageRevision.def()
     }
 }
