+++ deepwell/src/models/page_vote.rs
@@ -1,0 +1,54 @@
+//! SeaORM Entity. Generated by sea-orm-codegen 0.6.0
+
+use sea_orm::entity::prelude::*;
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
+#[sea_orm(table_name = "page_vote")]
+pub struct Model {
+    #[sea_orm(primary_key)]
+    pub page_vote_id: i64,
+    pub created_at: DateTimeWithTimeZone,
+    pub deleted_at: Option<DateTimeWithTimeZone>,
+    pub disabled_at: Option<DateTimeWithTimeZone>,
+    pub disabled_by: Option<i64>,
+    pub page_id: Option<i64>,
+    pub user_id: Option<i64>,
+    pub value: i16,
+}
+
+#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+pub enum Relation {
+    #[sea_orm(
+        belongs_to = "super::users::Entity",
+        from = "Column::DisabledBy",
+        to = "super::users::Column::Id",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    Users2,
+    #[sea_orm(
+        belongs_to = "super::page::Entity",
+        from = "Column::PageId",
+        to = "super::page::Column::PageId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    Page,
+    #[sea_orm(
+        belongs_to = "super::users::Entity",
+        from = "Column::UserId",
+        to = "super::users::Column::Id",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    Users1,
+}
+
+impl Related<super::page::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::Page.def()
+    }
+}
+
+impl ActiveModelBehavior for ActiveModel {}
