+++ deepwell/src/models/prelude.rs
@@ -1,4 +1,4 @@
-//! SeaORM Entity. Generated by sea-orm-codegen 0.5.0
+//! SeaORM Entity. Generated by sea-orm-codegen 0.6.0
 
 pub use super::file::Entity as File;
 pub use super::forum_category::Entity as ForumCategory;
@@ -10,8 +10,9 @@ pub use super::page_category::Entity as PageCategory;
 pub use super::page_connection::Entity as PageConnection;
 pub use super::page_connection_missing::Entity as PageConnectionMissing;
 pub use super::page_link::Entity as PageLink;
-pub use super::page_rate_vote::Entity as PageRateVote;
+pub use super::page_parent::Entity as PageParent;
 pub use super::page_revision::Entity as PageRevision;
+pub use super::page_vote::Entity as PageVote;
 pub use super::settings::Entity as Settings;
 pub use super::site::Entity as Site;
 pub use super::site_settings::Entity as SiteSettings;
